Simplify Windows compilation with Visual Studio
 * Restructured CMakeLists.txt for easier building on Windows
        ! I still cannot get the current build to compile on Windows because of
          weird linker errors but all other build steps seem fine
 * Removed the ability to use the system version of Bullet since it is often
   out of date
 * Rewrote Windows compilation instructions in README.md for Visual Studio
 + Added BEE_VERSION defines to bee/defines.hpp
 * Renamed the ProgramFlags struct to simply ProgramFlag
 + Added the E_FLAGARG enum as a replacement for getopt's argument types
 * Replaced getopt code in bee/init/programflags.cpp with custom argument
   parsing code
 * Moved send_packet() and recv_packet() to the net namespace
 * Made InstanceSort use a constant comparator
 * Fixed some issues and line endings in bee/util/platform.cpp
 + Added more undefs to bee/util/windefine.hpp
 * Updated build.sh with dependency building for Bullet

+++ bee/core/instance.cpp
@@ -732,7 +732,7 @@ namespace bee {
 	}
 	std::pair<int,int> Instance::get_snapped() const {
 		if (get_sprite() == nullptr) {
-			return std::make_pair(get_x(), get_y());
+			return std::make_pair(static_cast<int>(get_x()), static_cast<int>(get_y()));
 		}
 		return get_snapped(get_sprite()->get_width(), get_sprite()->get_height());
 	}
